MODFISTO-86 Update encumbrance doesn't work as expected
- renamed EncumbranceHandler to OrderTransactionsHandler
- renamed PaymentCreditHandler to InvoiceTransactionHandler
- moved code related to encumbrance update to InvoiceTransactionHandler
